refactor: Rename targets-view to target-cards-view
