+++ pkg/webui/ui/src/components/ErrorMessage.tsx
@@ -1,5 +1,5 @@
 import { Box, Typography, useTheme } from "@mui/material";
-import { CardPaper } from "./targets-view/Card";
+import { CardPaper } from "./target-cards-view/Card";
 
 export interface ErrorMessageProps {
     children?: React.ReactNode;
