+++ pkg/webui/ui/src/components/target-cards-view/ExpandableCard.tsx
