+++ pkg/webui/ui/src/components/target-cards-view/ProjectCard.tsx
@@ -5,7 +5,7 @@ import { ProjectIcon } from "../../icons/Icons";
 import { ProjectSummary } from "../../project-summaries";
 import { CardTemplate, cardWidth, projectCardMinHeight } from "./Card";
 
-export const ProjectItem = React.memo((props: { ps: ProjectSummary }) => {
+export const ProjectCard = React.memo((props: { ps: ProjectSummary }) => {
     const name = getLastPathElement(props.ps.project.gitRepoKey)
     const subDir = props.ps.project.subDir
 
@@ -35,4 +35,4 @@ export const ProjectItem = React.memo((props: { ps: ProjectSummary }) => {
     />;
 });
 
-ProjectItem.displayName = 'ProjectItem';
+ProjectCard.displayName = 'ProjectItem';
